update catalog classes and unit tests to use default Rv=3.1
we have moved to setting Rv=3.1 as a default_column in InstanceCatalog
classes.  This commit updates the GalSimCatalog classes to reflect that.
It also updates unit tests to reflect that reality (in some cases,
adding a getter that artificially sets Av=0.1 so that counts generated
in images match expected counts calculated by sims_photUtils without
having to calculate actual EBV values for every object in the fake catalog)