update examples to use CameraWrapper
I was not able to test this commit, as an update to fatboy has
caused CatSim database queries to run very slowly today.  I am
going to go ahead and merge in the interest of getting this work
done.  I will revisit the examples and make sure that they run
after fatboy is back to nominal performance.