update checkpointing
this required expanding the assertions in the checkpointing
unit test.  In the old version, the two images shared the same
WCS object in memory.  That is no longer the case.  Even
though the WCSs are equivalent, wcs1==wcs2 evaluates to False
because __eq__() has not been implemented for the relevant class.
The unit test now test the contents of the WCSs to make sure they
are equivalent